+++ api/server/routes/admin/auth.js
@@ -9,6 +9,7 @@ const { requireCapability } = require('~/server/middleware/roles/capabilities');
 const { createOAuthHandler } = require('~/server/controllers/auth/oauth');
 const { findBalanceByUser, upsertBalanceFields } = require('~/models');
 const { getAppConfig } = require('~/server/services/Config');
+const { storeAndStripChallenge } = require('~/server/utils/adminPkce');
 const getLogStores = require('~/cache/getLogStores');
 const { getOpenIdConfig } = require('~/strategies');
 const middleware = require('~/server/middleware');
@@ -73,66 +74,12 @@ router.get('/oauth/openid/check', (req, res) => {
   res.status(200).json({ message: 'OpenID check successful' });
 });
 
-/** PKCE challenge cache TTL: 5 minutes (enough for user to authenticate with IdP) */
-const PKCE_CHALLENGE_TTL = 5 * 60 * 1000;
-/** Regex pattern for valid PKCE challenges: 64 hex characters (SHA-256 hex digest) */
-const PKCE_CHALLENGE_PATTERN = /^[a-f0-9]{64}$/;
-
-/**
- * Strips `code_challenge` from the request query and URL strings.
- *
- * openid-client v6's Passport Strategy uses `currentUrl.searchParams.size === 0`
- * to distinguish an initial authorization request from an OAuth callback.
- * The admin-panel-specific `code_challenge` query parameter would cause the
- * strategy to misclassify the request as a callback and return 401.
- *
- * Applied defensively to all providers to ensure the admin-panel-private
- * `code_challenge` parameter never reaches any Passport strategy.
- * @param {import('express').Request} req
- */
-function stripCodeChallenge(req) {
-  delete req.query.code_challenge;
-  const strip = (url) =>
-    url.replace(/\?code_challenge=[^&]*&/, '?').replace(/[?&]code_challenge=[^&]*/, '');
-  req.originalUrl = strip(req.originalUrl);
-  req.url = strip(req.url);
-}
-
 /**
  * Generates a random hex state string for OAuth flows.
  * @returns {string} A 32-byte random hex string.
  */
 function generateState() {
   return crypto.randomBytes(32).toString('hex');
-}
-
-/**
- * Stores the admin-panel PKCE challenge in cache, then strips `code_challenge`
- * from the request so it doesn't interfere with the Passport strategy.
- *
- * Must be called before `passport.authenticate()` — the two operations are
- * logically atomic: read the challenge from the query, persist it, then remove
- * the parameter from the request URL.
- * @param {import('express').Request} req
- * @param {string} state - The OAuth state value (cache key).
- * @param {string} provider - Provider name for logging.
- * @returns {Promise<boolean>} True if stored (or no challenge provided); false on cache failure.
- */
-async function storeAndStripChallenge(req, state, provider) {
-  const { code_challenge: codeChallenge } = req.query;
-  if (typeof codeChallenge !== 'string' || !PKCE_CHALLENGE_PATTERN.test(codeChallenge)) {
-    stripCodeChallenge(req);
-    return true;
-  }
-  try {
-    const cache = getLogStores(CacheKeys.ADMIN_OAUTH_EXCHANGE);
-    await cache.set(`pkce:${state}`, codeChallenge, PKCE_CHALLENGE_TTL);
-    stripCodeChallenge(req);
-    return true;
-  } catch (err) {
-    logger.error(`[admin/oauth/${provider}] Failed to store PKCE challenge:`, err);
-    return false;
-  }
 }
 
 /**
@@ -498,4 +445,3 @@ router.post('/oauth/exchange', middleware.loginLimiter, async (req, res) => {
 });
 
 module.exports = router;
-module.exports._test = { stripCodeChallenge, storeAndStripChallenge };
